+++ tests/crossapp_retire_child/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class CrossappRetireChildConfig(AppConfig):
+    name = 'tests.crossapp_retire_child'
+    label = 'crossapp_retire_child'