+++ tests/crossapp_retire_owner/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class CrossappRetireOwnerConfig(AppConfig):
+    name = 'tests.crossapp_retire_owner'
+    label = 'crossapp_retire_owner'